Add pwd(1), getcwd(2); do code sharing of string.h
A lot of this was preparation for potentially porting some software, but this is till in the works. For now, we have one true string.h implementation: lib/string.c. This means that code is not duplicated across the kernel and userspace as much. In one particular instance, this actually led to a bug in sh(1), where the null terminator was not included in a strncpy. Unifying the implementations helped to point this out.

Additionally, as is standard, this commit passes all 4 build types and usertests(1).

+++ kernel/drivers/memide.c
@@ -6,18 +6,18 @@
 #include <buf.h>
 #include "fs.h"
 #include "console.h"
-#include "kernel_string.h"
+#include <string.h>
 
 extern uint8_t _binary_fs_img_start[], _binary_fs_img_size[];
 
-static int disksize;
+static size_t disksize;
 static uint8_t *memdisk;
 
 void
 ideinit(void)
 {
 	memdisk = _binary_fs_img_start;
-	disksize = (uint32_t)_binary_fs_img_size / BSIZE;
+	disksize = (size_t)_binary_fs_img_size / BSIZE;
 }
 
 // Interrupt handler.
